refactor button & iconbutton
