+++ packages/components/src/IconButton/index.tsx
@@ -1,93 +1,82 @@
-import { useContext } from 'react';
-
-import { styled, withStaticProperties } from 'tamagui';
-
-import { ButtonContext, ButtonFrame } from '../Button';
+import {
+  ButtonFrame,
+  type ButtonProps,
+  getSharedButtonStyles,
+} from '../Button';
 import { Icon } from '../Icon';
 import { Spinner } from '../Spinner';
 import { Stack } from '../Stack';
 
-import type { IconProps } from '../Icon';
-import type { ColorTokens } from 'tamagui';
+import type { ICON_NAMES } from '../Icon';
 
-const iconColorMapping: Record<string, ColorTokens> = {
-  primary: '$iconInverse',
-  secondary: '$icon',
-  tertiary: '$iconSubdued',
-  destructive: '$iconOnColor',
+interface IconButtonProps
+  extends Omit<ButtonProps, 'iconAfter' | 'children' | 'icon'> {
+  icon: ICON_NAMES;
+}
+
+const getSizeStyles = (size: ButtonProps['size']) => {
+  const sizes = {
+    small: {
+      p: '$1',
+      negativeMargin: '$-1',
+    },
+    medium: {
+      p: '$1.5',
+      negativeMargin: '$-1.5',
+    },
+    large: {
+      p: '$3',
+      negativeMargin: '$-3',
+    },
+  };
+
+  return sizes[size || 'medium'];
 };
 
-export const IconButtonIcon = (props: IconProps) => {
-  const { size, buttonVariant } = useContext(ButtonContext);
+export const IconButton = (props: IconButtonProps) => {
+  const {
+    disabled,
+    loading,
+    icon,
+    size,
+    variant = 'secondary',
+    ...rest
+  } = props;
+
+  const { p, negativeMargin } = getSizeStyles(size);
+
+  const { sharedFrameStyles, iconColor } = getSharedButtonStyles({
+    disabled,
+    loading,
+    variant,
+  });
 
   return (
-    <Icon
-      color={iconColorMapping[buttonVariant]}
-      size={size === 'small' ? '$5' : '$6'}
-      {...props}
-    />
-  );
-};
-
-export const IconButtonSpinner = () => {
-  const { size, buttonVariant } = useContext(ButtonContext);
-
-  return (
-    <Stack padding={size !== 'small' ? '$0.5' : '$0'}>
-      <Spinner color={iconColorMapping[buttonVariant]} />
-    </Stack>
-  );
-};
-
-const IconButtonStack = ButtonFrame.styleable((props, ref) => {
-  const { name, spinning, children } = props;
-
-  let childrenRender = children;
-  if (spinning) {
-    childrenRender = <IconButtonSpinner />;
-  } else if (name) {
-    childrenRender = <IconButtonIcon name={name} />;
-  }
-  return (
-    <ButtonFrame ref={ref} {...props}>
-      {childrenRender}
+    <ButtonFrame
+      p={p}
+      borderRadius="$full"
+      disabled={disabled || loading}
+      {...(variant === 'tertiary' && {
+        m: negativeMargin,
+      })}
+      {...sharedFrameStyles}
+      {...rest}
+    >
+      {loading ? (
+        <Stack
+          {...(size !== 'small' && {
+            p: '$0.5',
+          })}
+        >
+          <Spinner color={iconColor} size="small" />
+        </Stack>
+      ) : (
+        <Icon
+          color={iconColor}
+          name={icon}
+          size={size === 'small' ? '$5' : '$6'}
+        />
+      )}
     </ButtonFrame>
   );
-});
-
-const IconButtonFrame = styled(IconButtonStack, {
-  name: 'IconButton',
-  context: ButtonContext,
-  borderRadius: '$full',
-
-  variants: {
-    size: {
-      small: {
-        padding: '$1',
-      },
-      medium: {
-        padding: '$1.5',
-      },
-      large: {
-        padding: '$3',
-        borderRadius: '$full',
-      },
-    },
-
-    buttonVariant: {
-      secondary: {},
-      tertiary: {
-        padding: '$2',
-        margin: '$-2',
-      },
-      primary: {},
-      destructive: {},
-    },
-  } as const,
-});
-
-export const IconButton = withStaticProperties(IconButtonFrame, {
-  Props: ButtonContext.Provider,
-  Icon: IconButtonIcon,
-  Spinner: IconButtonSpinner,
-});
+};
