+++ packages/kit/src/routes/Root/Tab/Home/TabHomeStack1.tsx
@@ -1,4 +1,4 @@
-import { NewButton, YStack } from '@onekeyhq/components';
+import { Button, YStack } from '@onekeyhq/components';
 import type { PageNavigationProp } from '@onekeyhq/components/src/Navigation';
 
 import useAppNavigation from '../../../../hooks/useAppNavigation';
@@ -12,20 +12,20 @@ const TabHomeStack1 = () => {
 
   return (
     <YStack>
-      <NewButton
+      <Button
         onPress={() => {
           navigation.pop();
         }}
       >
         上一页
-      </NewButton>
-      <NewButton
+      </Button>
+      <Button
         onPress={() => {
           navigation.push(TabHomeRoutes.TabHomeStack2);
         }}
       >
         下一页
-      </NewButton>
+      </Button>
     </YStack>
   );
 };
