+++ packages/kit/src/routes/Root/Tab/Home/TabHomeStack2.tsx
@@ -1,4 +1,4 @@
-import { NewButton, YStack } from '@onekeyhq/components';
+import { Button, YStack } from '@onekeyhq/components';
 import type { PageNavigationProp } from '@onekeyhq/components/src/Navigation';
 
 import useAppNavigation from '../../../../hooks/useAppNavigation';
@@ -12,20 +12,20 @@ const TabHomeStack2 = () => {
 
   return (
     <YStack>
-      <NewButton
+      <Button
         onPress={() => {
           navigation.pop();
         }}
       >
         上一页
-      </NewButton>
-      <NewButton
+      </Button>
+      <Button
         onPress={() => {
           navigation.push(TabHomeRoutes.TabHome);
         }}
       >
         回首页
-      </NewButton>
+      </Button>
     </YStack>
   );
 };
