+++ packages/kit/src/views/Components/stories/IconButton.tsx
@@ -1,87 +1,62 @@
-import { useState } from 'react';
-
-import { XStack } from 'tamagui';
-
-import { IconButton } from '@onekeyhq/components';
+import { IconButton, Stack } from '@onekeyhq/components';
 
 import { Layout } from './utils/Layout';
 
-const ActionIconButton = () => {
-  const [spinning, setSpinning] = useState(false);
-  const [disabled, setDisabled] = useState(false);
-  return (
-    <IconButton
-      buttonVariant="secondary"
-      name="PlaceholderOutline"
-      spinning={spinning}
-      disabled={disabled}
-      onPress={() => {
-        setSpinning(true);
-        setDisabled(true);
-        setTimeout(() => {
-          setSpinning(false);
-          setDisabled(false);
-        }, 2000);
-      }}
-    />
-  );
-};
-
 const IconButtonGallery = () => (
   <Layout
-    description="对操作结果的反馈，无需用户操作即可自行消失"
-    suggestions={[
-      '使用 Toast 显示简约明确的信息反馈',
-      '用户点击或触摸 Toast 内容时，浮层将会停留在页面上',
-      'Toast 显示的文本应少于 20 字',
-      '不建议使用 Toast 显示过长的报错信息',
-    ]}
-    boundaryConditions={[
-      'Toast 永远拥有最高层级的浮层',
-      'Toast 组件能显示的最长文本内容为三排，超出三排将会缩略',
-      '界面中只会存在一个 Toast 示例，后触发的 Toast 信息会覆盖前一条 Toast 信息',
-    ]}
+    description=".."
+    suggestions={['...']}
+    boundaryConditions={['...']}
     elements={[
       {
-        title: 'Variants',
+        title: 'Varaints',
         element: (
-          <XStack space="$2" alignItems="center">
-            <IconButton buttonVariant="secondary" name="PlaceholderOutline" />
-            <IconButton buttonVariant="primary" name="PlaceholderOutline" />
-            <IconButton buttonVariant="tertiary" name="PlaceholderOutline" />
-            <IconButton buttonVariant="destructive" name="PlaceholderOutline" />
-          </XStack>
+          <Stack flexDirection="row" space="$4" alignItems="center">
+            <IconButton icon="PlaceholderOutline" />
+            <IconButton variant="primary" icon="PlaceholderOutline" />
+            <IconButton variant="destructive" icon="PlaceholderOutline" />
+            <IconButton variant="tertiary" icon="PlaceholderOutline" />
+          </Stack>
         ),
       },
       {
         title: 'Sizes',
         element: (
-          <XStack space="$2" alignItems="center">
-            <IconButton
-              buttonVariant="secondary"
-              size="large"
-              name="PlaceholderOutline"
-            />
-            <IconButton
-              buttonVariant="secondary"
-              size="medium"
-              name="PlaceholderOutline"
-            />
-            <IconButton
-              buttonVariant="secondary"
-              size="small"
-              name="PlaceholderOutline"
-            />
-          </XStack>
+          <Stack flexDirection="row" space="$4" alignItems="center">
+            <IconButton icon="PlaceholderOutline" />
+            <IconButton size="small" icon="PlaceholderOutline" />
+            <IconButton size="large" icon="PlaceholderOutline" />
+          </Stack>
         ),
       },
       {
-        title: 'Disabled & Loading',
+        title: 'Disabled',
         element: (
-          <XStack space="$2" alignItems="center">
-            <IconButton disabled name="PlaceholderOutline" />
-            <ActionIconButton />
-          </XStack>
+          <Stack flexDirection="row" space="$4">
+            <IconButton disabled icon="PlaceholderOutline" />
+            <IconButton disabled variant="primary" icon="PlaceholderOutline" />
+            <IconButton
+              disabled
+              variant="destructive"
+              icon="PlaceholderOutline"
+            />
+            <IconButton disabled variant="tertiary" icon="PlaceholderOutline" />
+          </Stack>
+        ),
+      },
+      {
+        title: 'Loading',
+        element: (
+          <Stack flexDirection="row" space="$4">
+            <IconButton loading icon="PlaceholderOutline" />
+            <IconButton loading variant="primary" icon="PlaceholderOutline" />
+            <IconButton
+              loading
+              variant="destructive"
+              icon="PlaceholderOutline"
+            />
+            <IconButton loading variant="tertiary" icon="PlaceholderOutline" />
+          </Stack>
         ),
       },
     ]}
