+++ packages/kit/src/views/Tab/Home/HomePageTabs/HeaderView.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useMemo, useState } from 'react';
 
-import { NewButton, Stack, Text } from '@onekeyhq/components';
+import { Button, Stack, Text } from '@onekeyhq/components';
 import type { PageNavigationProp } from '@onekeyhq/components/src/Navigation';
 
 import useAppNavigation from '../../../../hooks/useAppNavigation';
@@ -39,9 +39,9 @@ export default function HomePageHeaderView({
         <Text>Header View Simple</Text>
         <Text>{`Header Height ${headerHighMode.toString()}`}</Text>
         {headerHighMode && <Text py="$10">Very high</Text>}
-        <NewButton onPress={headerHeightCall}>切换高度</NewButton>
-        <NewButton onPress={switchDemoVisibleCall}>切换 Demo3 显示</NewButton>
-        <NewButton onPress={onNextPageCall}>下一页</NewButton>
+        <Button onPress={headerHeightCall}>切换高度</Button>
+        <Button onPress={switchDemoVisibleCall}>切换 Demo3 显示</Button>
+        <Button onPress={onNextPageCall}>下一页</Button>
       </Stack>
     ),
     [headerHighMode, headerHeightCall, onNextPageCall, switchDemoVisibleCall],
